api_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
deltalake==0.15.0
pandas==2.2.0
pydantic-settings==2.1.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1